test: Fix the instability issue in ASM E2E tests, as well as the failure issue in asm-cost-center.e2e (#17487)
