+++ projects/storefrontapp-e2e-cypress/cypress/e2e/b2b/regression/asm/asm-cost-center.e2e.cy.ts
@@ -36,7 +36,7 @@ context('B2B - ASM Account Checkout', () => {
     checkout.visitHomePage('asm=true');
     cy.get('cx-asm-main-ui').should('exist');
     cy.get('cx-asm-main-ui').should('be.visible');
-    asm.agentLogin('asagent', 'pw4all');
+    asm.agentLogin('brandon.leclair@acme.com', 'pw4all');
     cy.log('--> Agent emulate customer');
     asm.startCustomerEmulation(customer, true);
 
@@ -48,47 +48,21 @@ context('B2B - ASM Account Checkout', () => {
     cy.intercept('PUT', '*costcenter?costCenterId=*').as('costCenterReq');
 
     cy.get('cx-cost-center').within(() => {
-      cy.get('select').select(invalid_cost_center);
-    });
-
-    cy.wait('@costCenterReq').its('response.statusCode').should('eq', 400);
-    alerts.getErrorAlert().contains('Invalid cost center.');
-    alerts.getErrorAlert().then((alert) => {
-      cy.wrap(alert).within(() => {
-        cy.get('button').click();
+      cy.get('select').then((select) => {
+        cy.get('select').select(valid_cost_center);
+        if (select.find(`option:contains("${invalid_cost_center}")`).length) {
+          cy.get('select').select(invalid_cost_center);
+          cy.wait('@costCenterReq')
+            .its('response.statusCode')
+            .should('eq', 400);
+          alerts.getErrorAlert().contains('Invalid cost center.');
+          alerts.getErrorAlert().then((alert) => {
+            cy.wrap(alert).within(() => {
+              cy.get('button').click();
+            });
+          });
+        }
       });
     });
-    asm.agentSignOut();
-  });
-
-  it('should not show error on valid cost center', () => {
-    cy.log('--> Agent logging in');
-    checkout.visitHomePage('asm=true');
-    cy.get('cx-asm-main-ui').should('exist');
-    cy.get('cx-asm-main-ui').should('be.visible');
-    asm.agentLogin('asagent', 'pw4all');
-    cy.log('--> Agent emulate customer');
-    asm.startCustomerEmulation(customer, true);
-
-    b2bCheckout.addB2bProductToCartAndCheckout();
-    cy.get('cx-payment-type').within(() => {
-      cy.findByText('Account').click({ force: true });
-    });
-
-    cy.get('button.btn-primary').should('be.enabled').click({ force: true });
-
-    cy.intercept('PUT', '*costcenter?costCenterId=*').as('costCenterReq');
-    cy.get('cx-cost-center').within(() => {
-      cy.get('select').select(invalid_cost_center);
-      cy.wait('@costCenterReq');
-
-      cy.get('select').select(valid_cost_center);
-    });
-
-    cy.wait('@costCenterReq').its('response.statusCode').should('eq', 200);
-    alerts.getErrorAlert().should('not.exist');
-
-    cy.log('--> sign out');
-    asm.agentSignOut();
   });
 });
