+++ projects/storefrontapp-e2e-cypress/cypress/e2e/regression/asm/accessibility/tabbing-order.e2e.asm.e2e-flaky.cy.ts
@@ -4,12 +4,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { tabbingOrderConfig as config } from '../../helpers/accessibility/tabbing-order.config';
+import { tabbingOrderConfig as config } from '../../../../helpers/accessibility/tabbing-order.config';
 import {
   asmTabbingOrderWithCreateCustomerForm,
   asmTabbingOrderWithCustomerList,
   asmTabbingOrderWithSaveInactiveCartDialog,
-} from '../../helpers/accessibility/tabbing-order/asm';
+} from '../../../../helpers/accessibility/tabbing-order/asm';
 
 describe('Tabbing order for ASM', () => {
   before(() => {
