+++ projects/storefrontapp-e2e-cypress/cypress/e2e/regression/asm/asm.deeplink.core-e2e.cy.ts
@@ -4,31 +4,23 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import * as asm from '../../../../helpers/asm';
-import { login } from '../../../../helpers/auth-forms';
-import * as checkout from '../../../../helpers/checkout-flow';
-import { ELECTRONICS_BASESITE } from '../../../../helpers/checkout-flow';
-import * as customerTicketing from '../../../../helpers/customer-ticketing/customer-ticketing';
-import { getErrorAlert } from '../../../../helpers/global-message';
-import { signOutUser } from '../../../../helpers/login';
-import {
-  navigateToCategory,
-  waitForPage,
-} from '../../../../helpers/navigation';
-import { doPlaceOrder } from '../../../../helpers/order-history';
-import * as savedCart from '../../../../helpers/saved-cart';
-import { APPAREL_BASESITE } from '../../../../helpers/variants/apparel-checkout-flow';
-import * as sampleData from '../../../../sample-data/saved-cart';
-import { clearAllStorage } from '../../../../support/utils/clear-all-storage';
+import * as asm from '../../../helpers/asm';
+import { login } from '../../../helpers/auth-forms';
+import * as checkout from '../../../helpers/checkout-flow';
+import * as customerTicketing from '../../../helpers/customer-ticketing/customer-ticketing';
+import { signOutUser } from '../../../helpers/login';
+import { doPlaceOrder } from '../../../helpers/order-history';
+import * as savedCart from '../../../helpers/saved-cart';
+import * as sampleData from '../../../sample-data/saved-cart';
 
-import { getSampleUser } from '../../../../sample-data/checkout-flow';
+import { getSampleUser } from '../../../sample-data/checkout-flow';
 
 import {
   emulateCustomerPrepare,
   getCurrentCartIdAndAddProducts,
   getCustomerId,
   getInactiveCartIdAndAddProducts,
-} from '../../../../helpers/asm';
+} from '../../../helpers/asm';
 
 const agentToken = {
   userName: 'asagent',
@@ -36,9 +28,7 @@ const agentToken = {
 };
 
 context('Assisted Service Module', () => {
-  describe('Customer Support Agent - Emulation', () => {
-    asm.testCustomerEmulation();
-
+  describe('Customer Support Agent - Emulation&deeplink', () => {
     it('should emulate customer with deeplink before agent login (CXSPA-3113)', () => {
       const customer = getSampleUser();
       cy.log('--> Register new user');
@@ -52,9 +42,6 @@ context('Assisted Service Module', () => {
 
           cy.log('--> Should has assignCart');
           cy.get('.cx-asm-assignCart').should('exist');
-
-          cy.log('--> sign out and close ASM UI');
-          asm.agentSignOut();
         }
       );
     });
@@ -93,9 +80,6 @@ context('Assisted Service Module', () => {
             cy.get('.cx-asm-assignCart').should('exist');
 
             cy.url().should('contain', 'order/' + orderId);
-
-            cy.log('--> sign out and close ASM UI');
-            asm.agentSignOut();
           }
         );
       });
@@ -152,9 +136,6 @@ context('Assisted Service Module', () => {
             cy.get('.cx-asm-assignCart').should('exist');
 
             cy.url().should('contain', 'support-ticket/' + ticketId);
-
-            cy.log('--> sign out and close ASM UI');
-            asm.agentSignOut();
           });
         });
     });
@@ -202,9 +183,6 @@ context('Assisted Service Module', () => {
             cy.get('.cx-asm-assignCart').should('exist');
 
             cy.url().should('contain', 'saved-cart/' + savedCartId);
-
-            cy.log('--> sign out and close ASM UI');
-            asm.agentSignOut();
           }
         );
       });
@@ -225,9 +203,6 @@ context('Assisted Service Module', () => {
 
           cy.log('--> Should has assignCart');
           cy.get('.cx-asm-assignCart').should('exist');
-
-          cy.log('--> sign out and close ASM UI');
-          asm.agentSignOut();
         }
       );
     });
@@ -289,8 +264,6 @@ context('Assisted Service Module', () => {
           );
         });
       });
-      cy.log('--> sign out and close ASM UI');
-      asm.agentSignOut();
     });
 
     it('should not to switch emulated customer with deeplink when click cancel after agent login (CXSPA-3380)', () => {
@@ -346,9 +319,6 @@ context('Assisted Service Module', () => {
               'have.text',
               customerA.email
             );
-
-            cy.log('--> sign out and close ASM UI');
-            asm.agentSignOut();
           });
         }
       );
@@ -380,9 +350,6 @@ context('Assisted Service Module', () => {
           cy.visit('/assisted-service/emulate?customerId=notexist');
           cy.log('--> global error message should be display');
           cy.get('cx-global-message .alert-danger').should('be.visible');
-
-          cy.log('--> sign out and close ASM UI');
-          asm.agentSignOut();
         }
       );
     });
@@ -423,9 +390,6 @@ context('Assisted Service Module', () => {
             cy.get('cx-asm-main-ui').should('exist');
 
             cy.url().should('contain', 'order/' + orderId);
-
-            cy.log('--> sign out and close ASM UI');
-            asm.agentSignOut();
           }
         );
       });
@@ -485,9 +449,6 @@ context('Assisted Service Module', () => {
             cy.get('cx-asm-main-ui').should('be.visible');
 
             cy.url().should('contain', 'support-ticket/' + ticketId);
-
-            cy.log('--> sign out and close ASM UI');
-            asm.agentSignOut();
           });
         });
     });
@@ -535,9 +496,6 @@ context('Assisted Service Module', () => {
             cy.get('cx-asm-main-ui').should('be.visible');
 
             cy.url().should('contain', 'saved-cart/' + savedCartId);
-
-            cy.log('--> sign out and close ASM UI');
-            asm.agentSignOut();
           }
         );
       });
@@ -582,9 +540,6 @@ context('Assisted Service Module', () => {
             cy.get('cx-asm-main-ui').should('be.visible');
 
             cy.url().should('contain', '/cart');
-
-            cy.log('--> sign out and close ASM UI');
-            asm.agentSignOut();
           });
         }
       );
@@ -629,9 +584,6 @@ context('Assisted Service Module', () => {
             cy.get('cx-asm-main-ui').should('be.visible');
 
             cy.url().should('contain', '/cart');
-
-            cy.log('--> sign out and close ASM UI');
-            asm.agentSignOut();
           });
         }
       );
@@ -655,9 +607,6 @@ context('Assisted Service Module', () => {
 
           cy.log('--> Should not has assignCart');
           cy.get('.cx-asm-assignCart').should('not.exist');
-
-          cy.log('--> sign out and close ASM UI');
-          asm.agentSignOut();
         }
       );
     });
@@ -714,8 +663,6 @@ context('Assisted Service Module', () => {
             );
             cy.url().should('include', 'saved-cart');
             cy.url().should('include', inactiveCartId);
-            cy.log('--> sign out and close ASM UI');
-            asm.agentSignOut();
           }
         );
       });
@@ -763,118 +710,9 @@ context('Assisted Service Module', () => {
             );
             cy.get('button[id=asm-save-cart-dialog-btn]').should('be.disabled');
             cy.findByText(/Cancel/i).click();
-
-            cy.log('--> sign out and close ASM UI');
-            asm.agentSignOut();
           });
         }
       );
-    });
-
-    it('should checkout as customer', () => {
-      const customer = getSampleUser();
-
-      cy.log('--> Agent logging in');
-      checkout.visitHomePage('asm=true');
-      cy.get('cx-asm-main-ui').should('exist');
-      cy.get('cx-asm-main-ui').should('be.visible');
-
-      cy.log('--> Register user');
-      checkout.registerUser(false, customer);
-
-      asm.agentLogin('asagent', 'pw4all');
-
-      cy.log('--> Starting customer emulation');
-      asm.startCustomerEmulation(customer);
-
-      cy.log('--> Add product to cart and go to checkout');
-      checkout.goToCheapProductDetailsPage();
-      checkout.addCheapProductToCartAndBeginCheckoutForSignedInCustomer();
-
-      cy.log('--> Go through delivery form');
-      cy.contains('Continue').click();
-      checkout.fillAddressFormWithCheapProduct();
-
-      cy.log('--> Choose delivery method');
-      checkout.verifyDeliveryMethod();
-
-      cy.log('--> Fill payment form and continue');
-      checkout.fillPaymentForm();
-
-      cy.log('--> Place order');
-      checkout.placeOrderWithCheapProduct();
-
-      cy.log('--> sign out and close ASM UI');
-      asm.agentSignOut();
-    });
-  });
-
-  describe('When a customer session and an asm agent session are both active', () => {
-    let customer;
-
-    before(() => {
-      clearAllStorage();
-
-      cy.visit('/', { qs: { asm: true } });
-
-      customer = getSampleUser();
-      checkout.registerUser(false, customer);
-    });
-
-    it('Customer should not be able to login when there is an active CS agent session.', () => {
-      const loginPage = waitForPage('/login', 'getLoginPage');
-      cy.visit('/login?asm=true');
-      cy.wait(`@${loginPage}`);
-
-      asm.agentLogin('asagent', 'pw4all');
-      login(customer.email, customer.password);
-      getErrorAlert().should(
-        'contain',
-        'Cannot login as user when there is an active CS agent session. Please either emulate user or logout CS agent.'
-      );
-    });
-
-    // TODO(#9445): Add e2e test for this scenario
-    it.skip('agent login when user is logged in should start this user emulation', () => {});
-
-    // TODO(#9445): Add e2e test for this scenario
-    it.skip('agent logout when user was logged and emulated should restore the session', () => {});
-  });
-
-  describe('Apparel Site', () => {
-    before(() => {
-      Cypress.env('BASE_SITE', APPAREL_BASESITE);
-    });
-
-    after(() => {
-      Cypress.env('BASE_SITE', ELECTRONICS_BASESITE);
-    });
-
-    // This test only works if "sap-commerce-cloud-user-id" is added to the allowed headers of "corsfilter.commercewebservices.allowedHeaders" on the Commerce Cloud side. (CXSPA-1355)
-    it.skip("should fetch products in a category based on the emulated user's authentication", () => {
-      cy.cxConfig({
-        context: {
-          baseSite: ['apparel-uk-spa'],
-          currency: ['GBP'],
-        },
-      });
-
-      cy.visit('/', { qs: { asm: true } });
-
-      const customer = getSampleUser();
-      checkout.registerUser(false, customer);
-
-      asm.agentLogin('asagent', 'pw4all');
-
-      asm.startCustomerEmulation(customer);
-
-      navigateToCategory('Brands', 'brands', true);
-
-      cy.get('cx-product-list').should('exist');
-
-      navigateToCategory('Snow', 'snow', true);
-
-      cy.get('cx-product-list').should('exist');
     });
   });
 });
