+++ projects/storefrontapp-e2e-cypress/cypress/e2e/regression/asm/create-customer.e2e-flaky.cy.ts
@@ -4,11 +4,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import * as asm from '../../../../helpers/asm';
-import * as checkout from '../../../../helpers/checkout-flow';
-import { getSampleUser } from '../../../../sample-data/checkout-flow';
-import { myCompanyAdminUser } from '../../../../sample-data/shared-users';
-import { clearAllStorage } from '../../../../support/utils/clear-all-storage';
+import * as asm from '../../../helpers/asm';
+import * as checkout from '../../../helpers/checkout-flow';
+import { getSampleUser } from '../../../sample-data/checkout-flow';
+import { myCompanyAdminUser } from '../../../sample-data/shared-users';
+import { clearAllStorage } from '../../../support/utils/clear-all-storage';
 
 context('Assisted Service Module', () => {
   beforeEach(() => {
