+++ projects/storefrontapp-e2e-cypress/cypress/helpers/saved-cart.ts
@@ -247,12 +247,12 @@ export function saveActiveCart(verify = true) {
       // open modal to save the cart
 
       cy.get('cx-saved-cart-form-dialog').within(() => {
-        cy.get('[formcontrolname="name"]').type(
-          sampleData.savedActiveCartForm[0].name
-        );
-        cy.get('[formcontrolname="description"]').type(
-          sampleData.savedActiveCartForm[0].description
-        );
+        cy.get('[formcontrolname="name"]')
+          .clear()
+          .type(sampleData.savedActiveCartForm[0].name);
+        cy.get('[formcontrolname="description"]')
+          .clear()
+          .type(sampleData.savedActiveCartForm[0].description);
 
         cy.get('button[aria-label="Save"]').click();
       });
